StX Libraries Start
- XServo:
  Added documentation
  Added min and max variables
- XMotor:
  Added documentation
  Added the setPosition() method
  Added RPM monitorign and safety
- XOpMode:
  Added documentation.
  Added schedule object.
- XModule:
  Added documentation.
  Added schedule object.
  Made xGamepad updating more efficient.
- Button
  Added documentation
- StopWatch
  Added documentation
- XGamepad
  Added documentation
  More efficient update methods

+++ TeamCode/src/main/java/robotx/opmodes/autonomous/Park.java
@@ -22,7 +22,7 @@ public class Park extends LinearOpMode {
         autonMethods = new AutonMethods(this);
         autonMethods.init();
 
-        final double tileTime = AutonMethods.tileTime;
+        final int tileTime = AutonMethods.tileTime;
 
         autonMethods.start();
